acm-django/app/requirements.txt
django==1.8.5
mailsnake==1.6
djangorestframework==3.2.4
